+++ apps/fabric-website/src/pages/Components/FacepileComponentPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { FacepilePage } from '@uifabric/demo/lib/components/pages/FacepilePage';
+import { FacepilePage } from '@uifabric/fabric-website-resources/lib/components/pages/FacepilePage';
 import { PageHeader } from '../../components/PageHeader/PageHeader';
 import { ComponentPage } from '../../components/ComponentPage/ComponentPage';
 const pageStyles: any = require('../PageStyles.module.scss');