+++ apps/vr-tests/src/stories/MessageBar.stories.tsx
@@ -14,19 +14,21 @@ storiesOf('MessageBar', module)
   .addDecorator(FabricDecorator)
   .addDecorator(story => (
     <Screener
-      steps={ new Steps()
+      steps={new Steps()
         .snapshot('default', { cropTo: '.testWrapper' })
         .end()
       }
     >
-      { story() }
+      {story()}
     </Screener>
   ))
-  .add('Root', () => (<MessageBar>Info/default message bar. { link }</MessageBar>))
+  .add('Root', () => (<MessageBar>Info/default message bar. {link}</MessageBar>))
   .add('Root dismiss', () => (
-    <MessageBar onDismiss={ noop }>Info/default message bar. { link }</MessageBar>))
+    <MessageBar onDismiss={noop}>Info/default message bar. {link}</MessageBar>))
+  .add('Root dismiss single line', () => (
+    <MessageBar onDismiss={noop} isMultiline={false}>Info/default message bar. {link}</MessageBar>))
   .add('Root truncated', () => (
-    <MessageBar truncated={ true } isMultiline={ false }>Blocked lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi luctus, purus a lobortis tristique, odio augue pharetra metus, ac placerat nunc mi nec dui. Vestibulum aliquam et nunc semper scelerisque. Curabitur vitae orci nec quam condimentum porttitor et sed lacus. Vivamus ac efficitur leo. Cras faucibus mauris libero, ac placerat erat euismod et. Donec pulvinar commodo odio sit amet faucibus. In hac habitasse platea dictumst. Duis eu ante commodo, condimentum nibh pellentesque, laoreet enim. Fusce massa lorem, ultrices eu mi a, fermentum suscipit magna. Integer porta purus pulvinar, hendrerit felis eget, condimentum mauris. { link }</MessageBar>))
+    <MessageBar truncated={true} isMultiline={false}>Blocked lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi luctus, purus a lobortis tristique, odio augue pharetra metus, ac placerat nunc mi nec dui. Vestibulum aliquam et nunc semper scelerisque. Curabitur vitae orci nec quam condimentum porttitor et sed lacus. Vivamus ac efficitur leo. Cras faucibus mauris libero, ac placerat erat euismod et. Donec pulvinar commodo odio sit amet faucibus. In hac habitasse platea dictumst. Duis eu ante commodo, condimentum nibh pellentesque, laoreet enim. Fusce massa lorem, ultrices eu mi a, fermentum suscipit magna. Integer porta purus pulvinar, hendrerit felis eget, condimentum mauris. {link}</MessageBar>))
   .add('Root actions', () => (
     <MessageBar
       actions={
@@ -35,27 +37,61 @@ storiesOf('MessageBar', module)
           <MessageBarButton>No</MessageBarButton>
         </div>
       }
-    >Info/default message bar. { link }
+    >Info/default message bar. {link}
     </MessageBar>))
-  .add('Root multiline', () => (<MessageBar isMultiline>Info/default message bar. { longText }</MessageBar>))
-  .add('Root overflow', () => (<MessageBar isMultiline={ false }>Info/default message bar. { longText }  </MessageBar>))
+  .add('Root actions single line', () => (
+    <MessageBar
+      isMultiline={false}
+      actions={
+        <div>
+          <MessageBarButton>Yes</MessageBarButton>
+          <MessageBarButton>No</MessageBarButton>
+        </div>
+      }
+    >Info/default message bar. {link}
+    </MessageBar>))
+  .add('Root dismiss and action', () => (
+    <MessageBar
+      onDismiss={noop}
+      actions={
+        <div>
+          <MessageBarButton>Yes</MessageBarButton>
+          <MessageBarButton>No</MessageBarButton>
+        </div>
+      }
+    >Info/default message bar. {link}
+    </MessageBar>))
+  .add('Root dismiss and action single line', () => (
+    <MessageBar
+      isMultiline={false}
+      onDismiss={noop}
+      actions={
+        <div>
+          <MessageBarButton>Yes</MessageBarButton>
+          <MessageBarButton>No</MessageBarButton>
+        </div>
+      }
+    >Info/default message bar. {link}
+    </MessageBar>))
+  .add('Root multiline', () => (<MessageBar isMultiline>Info/default message bar. {longText}</MessageBar>))
+  .add('Root overflow', () => (<MessageBar isMultiline={false}>Info/default message bar. {longText}  </MessageBar>))
   .add('Error', () => (
-    <MessageBar messageBarType={ MessageBarType.error }>
-      Error message bar. { link }
+    <MessageBar messageBarType={MessageBarType.error}>
+      Error message bar. {link}
     </MessageBar>))
   .add('Blocked', () => (
-    <MessageBar messageBarType={ MessageBarType.blocked }>
-      Blocked message bar. { link }
+    <MessageBar messageBarType={MessageBarType.blocked}>
+      Blocked message bar. {link}
     </MessageBar>))
   .add('Severe Warning', () => (
-    <MessageBar messageBarType={ MessageBarType.severeWarning }>
-      Severe Warning message bar. { link }
+    <MessageBar messageBarType={MessageBarType.severeWarning}>
+      Severe Warning message bar. {link}
     </MessageBar>))
   .add('Success', () => (
-    <MessageBar messageBarType={ MessageBarType.success }>
-      Success message bar. { link }
+    <MessageBar messageBarType={MessageBarType.success}>
+      Success message bar. {link}
     </MessageBar>))
   .add('Warning', () => (
-    <MessageBar messageBarType={ MessageBarType.warning }>
-      Warning message bar. { link }
+    <MessageBar messageBarType={MessageBarType.warning}>
+      Warning message bar. {link}
     </MessageBar>));