+++ packages/experiments/src/demo/AppDefinition.tsx
@@ -22,6 +22,12 @@ export const AppDefinition: IAppDefinition = {
           url: '#/examples/chiclet'
         },
         {
+          component: require<any>('../components/Card/CardPage').CardPage,
+          key: 'DashboardCard',
+          name: 'DashboardCard',
+          url: '#/examples/Card'
+        },
+        {
           component: require<any>('../components/FolderCover/FolderCoverPage').FolderCoverPage,
           key: 'FolderCover',
           name: 'FolderCover',
@@ -46,10 +52,28 @@ export const AppDefinition: IAppDefinition = {
           url: '#/examples/layoutgroup'
         },
         {
+          component: require<any>('../components/Recommendation/RecommendationPage').RecommendationPage,
+          key: 'Recommendation',
+          name: 'Recommendation',
+          url: '#/examples/recommendation'
+        },
+        {
           component: require<any>('../components/signals/SignalsPage').SignalsPage,
           key: 'Signals',
           name: 'Signals',
           url: '#/examples/signals'
+        },
+        {
+          component: require<any>('../components/Stack/StackPage').StackPage,
+          key: 'Stack',
+          name: 'Stack',
+          url: '#/examples/stack'
+        },
+        {
+          component: require<any>('../components/Text/TextPage').TextPage,
+          key: 'Text',
+          name: 'Text',
+          url: '#/examples/text'
         },
         {
           component: require<any>('../components/Tile/TilePage').TilePage,
@@ -89,9 +113,15 @@ export const AppDefinition: IAppDefinition = {
         },
         {
           component: require<any>('../components/fluent/theme/FluentThemePage').FluentThemePage,
-          key: 'FluentTheme',
-          name: 'Fluent Theme',
-          url: '#/examples/fluenttheme'
+          key: 'FluentColorTheme',
+          name: 'Fluent Color Theme',
+          url: '#/examples/fluent-color-theme'
+        },
+        {
+          component: require<any>('../components/fluent/styles/FluentStylesPage').FluentStylesPage,
+          key: 'FluentStyles',
+          name: 'Fluent Component Styles',
+          url: '#/examples/fluent-styles'
         }
       ]
     }