+++ packages/office-ui-fabric-react/jest.config.js
@@ -5,9 +5,8 @@ const config = createConfig({
   setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))],
 
   moduleNameMapper: {
-    // These mappings allow Jest to run snapshot tests against Example files
-    'office-ui-fabric-react/lib/(.*)': '<rootDir>/src/$1',
-    '@uifabric/example-app-base': '@uifabric/example-app-base/lib-commonjs/utilities/data'
+    // These mappings allow Jest to run snapshot tests against Example files.
+    'office-ui-fabric-react/lib/(.*)$': '<rootDir>/src/$1'
   },
 
   snapshotSerializers: [path.resolve(__dirname, './node_modules/@uifabric/jest-serializer-merge-styles')]